requirements.txt
astroid==1.5.3
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
google-api-core==0.1.2
google-auth==1.2.1
google-cloud-bigquery==0.28.0
google-cloud-core==0.28.0
google-cloud-pubsub==0.29.4
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.7.3
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.0.6
protobuf==3.5.0.post1
psutil==5.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pylint==1.7.5
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
tweepy==3.5.0
urllib3==1.22
wrapt==1.10.11
